+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 See this http://keepachangelog.com link for information on how we want this documented formatted.
 
+## v0.6.0
+
+#### Added
+
+* APPLICATION LOAD BALANACER SUPPORT!!! The new default. Use `rack: :api` option to handler for API Gateway support.
+
+#### Changed
+
+* Rake task `lamby:install` now defaults to `application_load_balancer`
+
 
 ## v0.5.1